[Chore] 게임 네비게이션 컨트롤러 상속 관계 구현
